1.19.4 + addon compat ish

+++ src/main/java/com/gempire/commands/impl/CommandGempireWander.java
@@ -27,7 +27,7 @@ public class CommandGempireWander extends CommandBase {
     }
 
     public int execute(CommandSourceStack source) throws CommandSyntaxException {
-        BlockPos pos = new BlockPos(source.getPosition().x, source.getPosition().y, source.getPosition().z);
+        BlockPos pos = new BlockPos((int) source.getPosition().x, (int) source.getPosition().y, (int) source.getPosition().z);
         AABB aabb = source.getPlayerOrException().getBoundingBox().inflate(12.0D);
         List<EntityGem> gems = source.getLevel().getEntitiesOfClass(EntityGem.class, aabb);
         for(EntityGem gem : gems){
