+++ src/main/java/com/gempire/tileentities/InjectorTE.java
@@ -256,7 +256,7 @@ public class InjectorTE extends RandomizableContainerBlockEntity implements IFlu
             invalid = true;
             BlockPos crystalPos = getBlockPos().above().above();
             BlockPos drillPos = getBlockPos().below();
-            BlockPos seedPos = this.getBlockPos().offset(new BlockPos(0, -Math.ceil(GemSeedTE.DRAIN_SIZE / 2) - 1 - 1, 0));
+            BlockPos seedPos = this.getBlockPos().offset(new BlockPos(0, (int) (-Math.ceil(GemSeedTE.DRAIN_SIZE / 2) - 1 - 1), 0));
             if (level.getBlockState(crystalPos).getBlock() instanceof PowerCrystalBlock) {
                 System.out.println(fluidValid());
                 if (this.level.getBlockState(seedPos) != Blocks.BEDROCK.defaultBlockState()) {
